Update wordpress monorepo (#19328)
This is a pretty major update from a package versioning perspective, as we
hadn't updated the wordpress monorepo for some time.

Of particular note is that the update to `@wordpress/browserslist-config` means
that the JS is no longer transpiled down to something that IE11 supports, so
users of that browser will likely encounter JS errors now.

Co-authored-by: Glen Davies <glen.davies@a8c.com>
Co-authored-by: Marcelo Serpa <81248+fullofcaffeine@users.noreply.github.com>
Co-authored-by: Brad Jorsch <brad.jorsch@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1078920099